Split the grid path into fetching, upstream's conventions and the readers
grids.py held all three at once, so the ODIM composite parser imported the
download engine and the ICON coordinate cache was a private global the suite
reset by name. Now: gridfiles fetches, odim and icon carry upstream's file
conventions, grids is the readers - the split the tabular path already had
between transfer, meteocsv and readers.

icon.clear_cache() replaces the reach into _ICON_COORDS_CACHE, and the ICON
tests move to their own module where the cache is an internal seam they own.

test_layering's _imports() only understood 'from foehn.X import y', so a
'from foehn import x' edge was invisible to every assertion in the file.
